Updated main sketch (Cat_Feeder.pde)
Added support for 'Summer Mode' and 'Winter Mode'. This allows you
to have 2 different set amounts of food dispensing. Our cats eat
more in the Winter, less in the Summer.

Note - when you power on the device, either LED A or LED B will blink 3
times:
LED A will blink 3 times when 'initializing' to show that system
is in Summer mode OR
LED B will blink 3 times when 'initializing' to show that system
is in Winter mode